+++ src/resources/views/categories/edit.blade.php
@@ -83,7 +83,7 @@
 
                                 <div class="form-group">
                                     {!! Form::label('long_description', 'Description') !!}
-                                    {!! Form::textarea('long_description', $category->long_description, array('class' => 'form-control')) !!}
+                                    {!! Form::textarea('long_description', $category->long_description, array('class' => 'form-control', 'style' => 'height:200px')) !!}
                                 </div>
                             </div>
                             @foreach(\Config::get('redminportal::translation') as $translation)
@@ -91,17 +91,29 @@
                                 <div class="tab-pane" id="lang-{{ $translation['lang'] }}">
                                     <div class="form-group">
                                         {!! Form::label($translation['lang'] . '_name', 'Title') !!}
-                                        {!! Form::text($translation['lang'] . '_name', $category_cn->name, array('class' => 'form-control')) !!}
+                                        @if ($translated)
+                                        {!! Form::text($translation['lang'] . '_name', (property_exists($translated, $translation['lang']) ? $translated->$translation['lang']->name : ''), array('class' => 'form-control')) !!}
+                                        @else
+                                        {!! Form::text($translation['lang'] . '_name', null, array('class' => 'form-control')) !!}
+                                        @endif
                                     </div>
 
                                     <div class="form-group">
                                         {!! Form::label($translation['lang'] . '_short_description', 'Summary') !!}
-                                        {!! Form::text($translation['lang'] . '_short_description', $category_cn->short_description, array('class' => 'form-control')) !!}
+                                        @if ($translated)
+                                        {!! Form::text($translation['lang'] . '_short_description', (property_exists($translated, $translation['lang']) ? $translated->$translation['lang']->short_description : ''), array('class' => 'form-control')) !!}
+                                        @else
+                                        {!! Form::text($translation['lang'] . '_short_description', null, array('class' => 'form-control')) !!}
+                                        @endif
                                     </div>
 
                                     <div class="form-group">
                                         {!! Form::label($translation['lang'] . '_long_description', 'Description') !!}
-                                        {!! Form::textarea($translation['lang'] . '_long_description', $category_cn->long_description, array('class' => 'form-control')) !!}
+                                        @if ($translated)
+                                        {!! Form::textarea($translation['lang'] . '_long_description', (property_exists($translated, $translation['lang']) ? $translated->$translation['lang']->long_description : ''), array('class' => 'form-control', 'style' => 'height:200px')) !!}
+                                        @else
+                                        {!! Form::textarea($translation['lang'] . '_long_description', null, array('class' => 'form-control', 'style' => 'height:200px')) !!}
+                                        @endif
                                     </div>
                                 </div>
                                 @endif