+++ src/resources/views/portfolios/create.blade.php
@@ -63,9 +63,10 @@
                 </div>
                 <div class="panel-body">
                     <ul class="nav nav-tabs" id="lang-selector">
-                        <li class="active"><a href="#lang-en">English</a></li>
-                        <li><a href="#lang-sc">中文</a></li>
-                    </ul>
+                           @foreach(\Config::get('redminportal::translation') as $translation)
+                           <li><a href="#lang-{{ $translation['lang'] }}">{{ $translation['name'] }}</a></li>
+                           @endforeach
+                        </ul>
                     <div class="tab-content">
                         <div class="tab-pane active" id="lang-en">
                             <div class="form-group">
@@ -80,25 +81,29 @@
 
                             <div class="form-group">
                                 {!! Form::label('long_description', 'Description') !!}
-                                {!! Form::textarea('long_description', null, array('class' => 'form-control')) !!}
+                                {!! Form::textarea('long_description', null, array('class' => 'form-control', 'style' => 'height:200px')) !!}
                             </div>
                         </div>
-                        <div class="tab-pane" id="lang-sc">
-                            <div class="form-group">
-                                {!! Form::label('cn_name', '标题') !!}
-                                {!! Form::text('cn_name', null, array('class' => 'form-control')) !!}
-                            </div>
+                        @foreach(\Config::get('redminportal::translation') as $translation)
+                            @if($translation['lang'] != 'en')
+                            <div class="tab-pane" id="lang-{{ $translation['lang'] }}">
+                                <div class="form-group">
+                                    {!! Form::label($translation['lang'] . '_name', 'Title') !!}
+                                    {!! Form::text($translation['lang'] . '_name', null, array('class' => 'form-control')) !!}
+                                </div>
 
-                            <div class="form-group">
-                                {!! Form::label('cn_short_description', '简介') !!}
-                                {!! Form::text('cn_short_description', null, array('class' => 'form-control')) !!}
-                            </div>
+                                <div class="form-group">
+                                    {!! Form::label($translation['lang'] . '_short_description', 'Summary') !!}
+                                    {!! Form::text($translation['lang'] . '_short_description', null, array('class' => 'form-control')) !!}
+                                </div>
 
-                            <div class="form-group">
-                                {!! Form::label('cn_long_description', '内容') !!}
-                                {!! Form::textarea('cn_long_description', null, array('class' => 'form-control')) !!}
+                                <div class="form-group">
+                                    {!! Form::label($translation['lang'] . '_long_description', 'Description') !!}
+                                    {!! Form::textarea($translation['lang'] . '_long_description', null, array('class' => 'form-control', 'style' => 'height:200px')) !!}
+                                </div>
                             </div>
-                        </div>
+                            @endif
+                        @endforeach
                     </div>
                 </div>
             </div>
@@ -112,6 +117,7 @@
     <script>
         !function ($) {
             $(function(){
+                $('#lang-selector li').first().addClass('active');
                 $('#lang-selector a').click(function (e) {
                     e.preventDefault();
                     $(this).tab('show');